+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## [153.0.2](https://github.com/Sage/carbon/compare/v153.0.1...v153.0.2) (2025-04-04)
+
+### Bug Fixes
+
+* **dialog-full-screen:** deprecate timeout and disableClose props in dialog-full-screen ([eca40d5](https://github.com/Sage/carbon/commit/eca40d58b128f9309986bb01449cae2b52a1b41e)), closes [#6182](https://github.com/Sage/carbon/issues/6182)
+* **dialog:** deprecate timeout prop inside dialog ([2448eef](https://github.com/Sage/carbon/commit/2448eef716d1015dd7691a64af7492046a0c3c6b)), closes [#6166](https://github.com/Sage/carbon/issues/6166)
+
+## [153.0.1](https://github.com/Sage/carbon/compare/v153.0.0...v153.0.1) (2025-04-02)
+
+### Bug Fixes
+
+* **button-bar:** ensure disabled buttons have correct styles ([553fb68](https://github.com/Sage/carbon/commit/553fb680cbf564d0bea5f39fb99351a9a661cb3e)), closes [#6292](https://github.com/Sage/carbon/issues/6292)
+
+## [153.0.0](https://github.com/Sage/carbon/compare/v152.7.0...v153.0.0) (2025-04-01)
+
+### ⚠ BREAKING CHANGES
+
+* **deps:** `styled-components` is a peer dependency so consumers will need to install the new
+version
+
+### Features
+
+* **deps:** upgrade styled-components to v5 ([c5ad5e3](https://github.com/Sage/carbon/commit/c5ad5e3ea90a59fd34e2e23c89f023ab7ed70e64))
+
 ## [152.7.0](https://github.com/Sage/carbon/compare/v152.6.0...v152.7.0) (2025-03-28)
 
 ### Features